# Backend/docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    email: support@filevault.com
    name: FileVault Support
  description: A secure file storage and sharing service with user authentication
    and admin management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FileVault API
  version: "1.0"
paths:
...
  /user/profile:
    get:
      consumes:
      - application/json
      description: Returns the current authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
  /user/storage:
    get:
      consumes:
      - application/json
      description: Returns the current user's storage usage and quota information
      produces:
      - application/json
      responses:
        "200":
          description: Storage information
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get storage information
      tags:
      - users
  /user/storage/statistics:
    get:
      consumes:
      - application/json
      description: Returns comprehensive storage statistics for the current user
      produces:
      - application/json
      responses:
        "200":
          description: Storage statistics
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get storage statistics
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"